major rewrite stage 2: kernel rewrite restructure the kernel, added more safety features, improved made ps2 driver actually work, added qemu  log, better code structure, and other small changes

+++ tools/build.sh
@@ -9,5 +9,7 @@ cp target/x86_64-unknown-uefi/release/bootx64.efi bin/efi/boot
 cp target/x86_64-unknown-none/release/kernelx64 bin/
 
 cp res/assets/fonts/kernel_terminal_font.psf bin/fonts/default.psf
-cp res/assets/config.toml bin/
+if ! test -f bin/config.toml; then
+    cp res/assets/config.toml bin/
+fi
 cp res/assets/themes/default.toml bin/themes
